+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Add support for forcing all relations to be undefined by default ([#23](https://github.com/daniel7grant/eslint-plugin-typeorm-typescript/pull/23))
 
+### Changed
+
+- Allow PrimaryColumn to support any type, as opposed to defaulting to integers ([#21](https://github.com/daniel7grant/eslint-plugin-typeorm-typescript/issues/21))
+- Add custom error message for bigints and decimals to make it clearer that it's not a false positive ([#22](https://github.com/daniel7grant/eslint-plugin-typeorm-typescript/issues/22))
+
 ## [0.5.0] - 2024-12-06
 
 ### Changed